[WIP] CM-116 refactor code with hexagonal architecture
