+++ .idea/.gitignore
@@ -70,3 +70,4 @@
 /php.xml
 /vcs.xml
 /git_toolbox_prj.xml
+/csv-editor.xml
