+++ clinicalResource/config/routes.yaml
@@ -6,8 +6,8 @@
 
 api:
     resource:
-        path: ../src/Controller/Api/v1
-        namespace: App\Controller\Api\v1
+        path: ../src/Infrastructure/Http/Api/V1
+        namespace: App\Infrastructure\Http\Api\V1
         prefix: '/v1'
     type: attribute
 
@@ -21,6 +21,6 @@ api_logout:
 
 web:
     resource:
-        path: ../src/Controller/Web
-        namespace: App\Controller\Web
+        path: ../src/Infrastructure/Http/Web
+        namespace: App\Infrastructure\Http\Web
     type: attribute
