+++ clinicalResource/src/Infrastructure/Form/RegistrationFormType.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Form;
+namespace App\Infrastructure\Form;
 
-use App\Entity\User;
+use App\Domain\Entity\User;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
