+++ clinicalResource/src/Infrastructure/Http/Api/V1/pruebasController.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Controller\Api\v1;
+namespace App\Infrastructure\Http\Api\V1;
 
-use Symfony\Component\HttpFoundation\JsonResponse;
-use \Symfony\Component\Routing\Attribute\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 #[Route('/api/v1/pruebas',  name: 'pruebas')]
