+++ clinicalResource/src/Infrastructure/Http/Web/Login/LogoutController.php
@@ -1,10 +1,8 @@
 <?php
 
-namespace App\Controller\Api;
+namespace App\Infrastructure\Http\Web\Login;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
